requirements.txt
google-generativeai
pypdf
streamlit
ddgs
uv
python-dotenv
